Conclusão de ficha Cadastro Individual e substituição de tipos primitivos nas entidades

+++ src/esaude/model/EsusAtividadeColetivaParticipantes.java
@@ -1,6 +1,7 @@
 package esaude.model; 
 
 import java.io.Serializable;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -9,7 +10,9 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
 import java.io.*;
+import java.math.BigDecimal;
 import java.util.*;
 
 @Entity 
@@ -31,7 +34,7 @@ public class EsusAtividadeColetivaParticipantes implements Serializable {
    private boolean avaliacaoAlterada;
 
    @Column(name="peso")
-   private double peso;
+   private BigDecimal peso;
 
    @Column(name="altura")
    private Long altura;
@@ -68,10 +71,10 @@ public class EsusAtividadeColetivaParticipantes implements Serializable {
       this.avaliacaoAlterada = avaliacaoAlterada;
    }
 
-   public double getPeso() { 
+   public BigDecimal getPeso() { 
       return peso;
    }
-   public void  setPeso(double peso) { 
+   public void  setPeso(BigDecimal peso) { 
       this.peso = peso;
    }
